Ordenación de archivos
He separado los archivos en py más pequeños, reformulada la función  mostrar inventario y  eliminado prompt (ahora  lo tiene dani)
