+++ m3/diccionarios.py
@@ -1,0 +1,22 @@
+#diccionario para el inventario
+inventory = {
+    'name': 'link',
+    'BloodMoon': '25',
+    'Hearts': {'current': 3, 'max': 3},
+    'food': {},
+    'weapons': []
+}
+
+# diccionario de contadores
+item_counters = {
+    'wooden_sword': 0,
+    'wooden_shield': 0,
+    'sword': 0,
+    'shield': 0,
+    'vegetable': 0,
+    'fish': 0,
+    'meat': 0,
+    'salad': 0,
+    'pescatarian': 0,
+    'roasted': 0
+}
